Refactored into separate graph class
